refactor: rename featured item model to avoid confusion with dv objects type collection
