+++ src/collection/domain/useCases/getCollectionFeaturedItems.ts
@@ -1,9 +1,9 @@
-import { CollectionFeaturedItem } from '../models/CollectionFeaturedItem'
+import { FeaturedItem } from '../models/FeaturedItem'
 import { CollectionRepository } from '../repositories/CollectionRepository'
 
 export async function getCollectionFeaturedItems(
   collectionRepository: CollectionRepository,
   collectionIdOrAlias?: number | string
-): Promise<CollectionFeaturedItem[]> {
+): Promise<FeaturedItem[]> {
   return collectionRepository.getFeaturedItems(collectionIdOrAlias)
 }
