+++ src/collection/domain/useCases/updateCollectionFeaturedItems.ts
@@ -1,4 +1,4 @@
-import { CollectionFeaturedItem } from '../models/CollectionFeaturedItem'
+import { FeaturedItem } from '../models/FeaturedItem'
 import { CollectionRepository } from '../repositories/CollectionRepository'
 import { CollectionFeaturedItemsDTO } from './DTOs/CollectionFeaturedItemsDTO'
 
@@ -6,6 +6,6 @@ export async function updateCollectionFeaturedItems(
   collectionRepository: CollectionRepository,
   featuredItems: CollectionFeaturedItemsDTO,
   collectionIdOrAlias: number | string
-): Promise<CollectionFeaturedItem[]> {
+): Promise<FeaturedItem[]> {
   return collectionRepository.updateFeaturedItems(collectionIdOrAlias, featuredItems)
 }
