+++ src/sections/collection/featured-items/FeaturedItems.tsx
@@ -4,7 +4,7 @@ import Skeleton, { SkeletonTheme } from 'react-loading-skeleton'
 import { Button, useTheme } from '@iqss/dataverse-design-system'
 import { ChevronLeft, ChevronRight } from 'react-bootstrap-icons'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
-import { FeaturedItemType } from '@/collection/domain/models/CollectionFeaturedItem'
+import { FeaturedItemType } from '@/collection/domain/models/FeaturedItem'
 import { useGetCollectionFeaturedItems } from '../useGetCollectionFeaturedItems'
 import { CustomFeaturedItemCard } from './custom-featured-item-card/CustomFeaturedItemCard'
 import { DvObjectFeaturedItemCard } from './dv-object-featured-item-card/DvObjectFeaturedItemCard'
