+++ src/sections/edit-collection-featured-items/featured-items-form/featured-item-field/FeaturedItemField.tsx
@@ -9,7 +9,7 @@ import { WriteError } from '@iqss/dataverse-client-javascript'
 import cn from 'classnames'
 import { Badge, Button, Col, Row, Tooltip } from '@iqss/dataverse-design-system'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
-import { FeaturedItemType } from '@/collection/domain/models/CollectionFeaturedItem'
+import { FeaturedItemType } from '@/collection/domain/models/FeaturedItem'
 import { BaseFormItem } from './base-form-item/BaseFormItem'
 import { DvObjectFormItem } from './dv-object-form-item/DvObjectFormItem'
 import { CustomFormItem } from './custom-form-item/CustomFormItem'
