+++ src/sections/featured-item/FeaturedItem.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Alert } from '@iqss/dataverse-design-system'
 import { CollectionRepository } from '@/collection/domain/repositories/CollectionRepository'
-import { CustomFeaturedItem } from '@/collection/domain/models/CollectionFeaturedItem'
+import { CustomFeaturedItem } from '@/collection/domain/models/FeaturedItem'
 import { useGetCollectionFeaturedItems } from '../collection/useGetCollectionFeaturedItems'
 import { AppLoader } from '../shared/layout/app-loader/AppLoader'
 import { useLoading } from '../loading/LoadingContext'
